requirements.txt
aiogram
python-decouple
gspread
pandas